style: 테스트 디렉터리 구조 변경
