+++ src/main/java/blackjack/constant/TrumpSuit.java
@@ -1,6 +1,7 @@
 package blackjack.constant;
 
 public enum TrumpSuit {
+
     SPADE("스페이드"),
     CLOVER("클로버"),
     HEART("하트"),
