+++ src/main/java/blackjack/gametable/gambler/PlayerName.java
@@ -1,4 +1,4 @@
-package blackjack.gambler;
+package blackjack.gametable.gambler;
 
 public class PlayerName {
