+++ src/test/java/blackjack/gametable/gambler/GamblerTest.java
@@ -1,0 +1,7 @@
+package blackjack.gametable.gambler;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class GamblerTest {
+
+}
