+++ src/test/java/blackjack/gametable/gambler/PlayerNameTest.java
@@ -1,6 +1,5 @@
-package domain;
+package blackjack.gametable.gambler;
 
-import blackjack.gambler.PlayerName;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
